refactor: move UserEntity to entities package
